+++ linter/linter.go
@@ -23,6 +23,7 @@ func Lint(r *resolver.Resolver) []Hint {
 	res = append(res, CheckAmbiguous(r)...)
 	res = append(res, CheckAssignees(r)...)
 	res = append(res, CheckNoContext(r)...)
+	res = append(res, CheckAnnotations(r)...)
 
 	return res
 }